change id type to Long

+++ src/main/java/com/delegrego/exemplo_spring_boot_2/controller/DepartamentoController.java
@@ -1,6 +1,5 @@
 package com.delegrego.exemplo_spring_boot_2.controller;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -74,7 +73,7 @@ public class DepartamentoController {
 	 * @return ResponseEntity com o departamento encontrado
 	 */
 	@GetMapping("/{id}")
-	public ResponseEntity<DepartamentoResponseDto> obterDepartamentoPorId(@PathVariable BigInteger id) {
+	public ResponseEntity<DepartamentoResponseDto> obterDepartamentoPorId(@PathVariable Long id) {
 		return ResponseEntity.status(HttpStatus.OK).body(servico.obterDepartamentoPorId(id));
 	}
 
@@ -97,7 +96,7 @@ public class DepartamentoController {
 	 * @return ResponseEntity com status apropriado
 	 */
 	@PutMapping("/{id}")
-	public ResponseEntity<DepartamentoEntity> atualizarDepartamento(@PathVariable BigInteger id,
+	public ResponseEntity<DepartamentoEntity> atualizarDepartamento(@PathVariable Long id,
 			@Valid @RequestBody DepartamentoRequestDto departamentoDto) {
 		return ResponseEntity.status(HttpStatus.OK).body(servico.atualizarDepartamento(id, departamentoDto));
 	}
@@ -109,7 +108,7 @@ public class DepartamentoController {
 	 * @return ResponseEntity com status apropriado
 	 */
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> deletarDepartamento(@PathVariable BigInteger id) {
+	public ResponseEntity<Void> deletarDepartamento(@PathVariable Long id) {
 		servico.deletarDepartamento(id);
 		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 	}
